src/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Arnau Rovira <arnaurovira81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.2"
fastapi = "^0.79.1"
prometheus-fastapi-instrumentator = "^5.9.1"
loguru = "^0.6.0"
httpx = "^0.23.1"
pytest = "^7.2.0"
requests = "^2.28.1"
pytidylib = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
